factor out command handling into own module
