+++ src/deltabot/deltabot.py
@@ -1,15 +1,11 @@
 # -*- coding: utf-8 -*-
 
-from collections import OrderedDict
-
 import deltachat as dc
 from deltachat import account_hookimpl
 from deltachat.tracker import ConfigureTracker
 
 from . import hookspec
-
-
-CMD_PREFIX = '/'
+from .commands import Commands
 
 
 class Filter():
@@ -17,39 +13,15 @@ class Filter():
         return False
 
 
-class CommandDef:
-    """ Definition of a '/COMMAND' with args. """
-    def __init__(self, cmd, short, long, func):
-        if cmd[0] != CMD_PREFIX:
-            raise ValueError("cmd {!r} must start with {!}".format(cmd, CMD_PREFIX))
-        self.cmd = cmd
-        self.long = long
-        self.short = short
-        self.func = func
-
-    def __eq__(self, c):
-        return c.__dict__ == self.__dict__
-
-
-class IncomingCommand:
-    """ incoming command request. """
-    def __init__(self, bot, cmd_def, payload, message):
-        self.bot = bot
-        self.cmd_def = cmd_def
-        self.payload = payload
-        self.message = message
-
-
-class CommandNotFound(LookupError):
-    """Command was not found. """
-
-
 class DeltaBot:
     def __init__(self, account, logger):
         self.account = account
         self._pm = hookspec.DeltaBotSpecs._make_plugin_manager()
         self.logger = logger
-        self._cmd_defs = OrderedDict()
+
+        #: commands subsystem for registering/executing commands
+        self.commands = Commands(self)
+
         self.filters = []
 
         # set some useful bot defaults
@@ -87,29 +59,6 @@ class DeltaBot:
     def list_plugins(self):
         """ return a dict name->deltabot plugin object mapping. """
         return dict(self._pm.list_name_plugin())
-
-    # =========================================================
-    # deltabot command API
-    # =========================================================
-    def register_command(self, name, func):
-        short, long = parse_command_docstring(func)
-        cmd_def = CommandDef(name, short=short, long=long, func=func)
-        if name in self._cmd_defs:
-            raise ValueError("command {!r} already registered".format(name))
-        self._cmd_defs[name] = cmd_def
-        self.logger.debug("registered new command {!r}".format(name))
-
-    def _process_command_message(self, message):
-        assert message.text.startswith(CMD_PREFIX)
-        parts = message.text.split(maxsplit=1)
-        cmd_name = parts.pop(0)
-        cmd_def = self._cmd_defs.get(cmd_name)
-        if cmd_def is None:
-            raise CommandNotFound("unknown {!r} command in message {!r}".format(
-                cmd_name, message))
-        payload = parts[0] if parts else ""
-        cmd = IncomingCommand(bot=self, cmd_def=cmd_def, payload=payload, message=message)
-        return cmd.cmd_def.func(cmd)
 
     def is_configured(self):
         return bool(self.account.is_configured())
@@ -165,17 +114,26 @@ class DeltaBot:
     @account_hookimpl
     def process_incoming_message(self, message):
         try:
+            # we always accept incoming messages to remove the need
+            # for bot authors to having to deal with deaddrop/contact
+            # request.  But we record whether it was one in case
+            # a bot author still wants to act on it. This way it's still possible
+            # to block or ignore an original contact request
             message.was_contact_request = message.chat.is_deaddrop()
             message.accept_sender_contact()
+
             self.logger.info("incoming message from {} id={} chat={} text={!r}".format(
                 message.get_sender_contact().addr,
                 message.id, message.chat.id, message.text[:50]))
-            if message.text and message.text.startswith(CMD_PREFIX):
-                res = self._process_command_message(message)
-                if res:
-                    reply = message.chat.send_text(res)
-                    self.logger.info("sending reply to chat={}: text={!r}".format(
-                        reply.chat.id, reply.text[:50]))
+            res = self.commands.process_command_message(message)
+            if res:
+                reply = message.chat.send_text(res)
+                self.logger.info("sending reply to chat={}: text={!r}".format(
+                    reply.chat.id, reply.text[:50]))
+            else:
+                self.logger.info("no action for message from {} id={}".format(
+                    message.get_sender_contact().addr, message.id
+                ))
         except Exception as ex:
             self.logger.exception(ex)
 
@@ -222,12 +180,3 @@ class DeltaBot:
 
     def is_group(self, chat):
         return chat.is_group()
-
-
-def parse_command_docstring(func):
-    description = func.__doc__
-    if not description:
-        raise ValueError("command {!r} needs to have a docstring".format(func))
-
-    lines = description.strip().split("\n")
-    return lines.pop(0), "\n".join(lines).strip()
